+++ src/mon/OSDMonitor.cc
@@ -2473,6 +2473,7 @@ void OSDMonitor::get_health(list<pair<health_status_t,string> >& summary,
 			 CEPH_OSDMAP_NOIN |
 			 CEPH_OSDMAP_NOOUT |
 			 CEPH_OSDMAP_NOBACKFILL |
+			 CEPH_OSDMAP_NOREBALANCE |
 			 CEPH_OSDMAP_NORECOVER |
 			 CEPH_OSDMAP_NOSCRUB |
 			 CEPH_OSDMAP_NODEEP_SCRUB |
@@ -5287,6 +5288,8 @@ bool OSDMonitor::prepare_command_impl(MMonCommand *m,
       return prepare_set_flag(m, CEPH_OSDMAP_NODEEP_SCRUB);
     else if (key == "notieragent")
       return prepare_set_flag(m, CEPH_OSDMAP_NOTIERAGENT);
+    else if (key == "norebalance")
+      return prepare_set_flag(m, CEPH_OSDMAP_NOREBALANCE);
     else {
       ss << "unrecognized flag '" << key << "'";
       err = -EINVAL;
@@ -5317,6 +5320,8 @@ bool OSDMonitor::prepare_command_impl(MMonCommand *m,
       return prepare_unset_flag(m, CEPH_OSDMAP_NODEEP_SCRUB);
     else if (key == "notieragent")
       return prepare_unset_flag(m, CEPH_OSDMAP_NOTIERAGENT);
+    else if (key == "norebalance")
+      return prepare_unset_flag(m, CEPH_OSDMAP_NOREBALANCE);
     else {
       ss << "unrecognized flag '" << key << "'";
       err = -EINVAL;